almost have the store loading actions
