+++ src/components/ExtendedForecast.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+
+export default class ExtendedForecast extends Component {
+  render() {
+    return (
+      <section id="forcast">
+        <article className="extended-weather-card">Current weather4</article>
+        <h1></h1>
+      </section>
+    )
+  }
+}
