+++ src/components/Home.js
@@ -15,8 +15,6 @@ export default class Home extends Component {
           <article className="weather-card">Current weather1</article>
           <article className="weather-card">Current weather2</article>
           <article className="weather-card">Current weather3</article>
-          <article className="weather-card">Current weather4</article>
-          <article className="weather-card">Current weather5</article>
           <Link to="/pinned-cities"><article className="add-city-card">+ Pin another city</article></Link>
         </section>
       </div>
