+++ src/index.js
@@ -1,24 +1,13 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { createStore, applyMiddleware, combineReducers } from 'redux';
 import { Provider } from 'react-redux';
 import { browserHistory } from 'react-router';
 import Routes from './routes';
-import thunk from 'redux-thunk';
-import createLogger from 'redux-logger';
-import reducer from './reducers';
+import reducer from './reducers/index';
+import store, { history } from './store';
+
 
 require("!style!css!sass!./styles/index.scss");
-// require('./styles/index.scss');
-
-const middleware = [thunk, createLogger];
-
-const store = createStore(
-  combineReducers({
-    weatherApp: reducer,
-  }),
-  applyMiddleware(...middleware)
-)
 
 render(
   <Provider store={store}>
