+++ src/reducers/GetWeather.js
@@ -1,0 +1,4 @@
+export const weather = (state = {}, action) => {
+  console.log(state, action);
+  return state;
+};
